refactoring to fix linter
