+++ src/App.tsx
@@ -1,5 +1,5 @@
 import Page from "./Page.tsx";
-import { AuthProvider } from "./contexts/AuthContext.tsx";
+import { AuthProvider } from "./contexts/Providers.tsx";
 
 export default function App() {
   return (
