--- src/contexts/AuthContext.tsx
@@ -1,67 +1,0 @@
-import { createContext, useEffect, useState } from "react";
-import { fetchOAuth } from "../Login.tsx";
-import {  OAuthResult} from "@huggingface/hub";
-
-interface AuthContextValue {
-  oAuthResult?: OAuthResult;
-  headers: Record<string, string>;
-  fetch: typeof window.fetch;
-}
-
-function createAuthFromOAuthResult(oAuthResult: OAuthResult): AuthContextValue {  
-  const headers = {
-    Authorization: `Bearer ${oAuthResult.accessToken}`,
-  };
-  return {
-    oAuthResult,
-    headers,
-    fetch: async (input, init) => {
-      if (init === undefined) init = {};
-      if (init.headers === undefined) init.headers = {};
-      init.headers = {
-        ...headers,
-        /// the Authorization header can be overridden by the caller
-        ...init.headers,
-      };
-      return window.fetch(input, init);
-    },
-  };
-}
-
-function createEmptyAuth(): AuthContextValue {
-  return {
-    headers: {},
-    fetch: window.fetch.bind(window),
-  };
-}
-
-// The AuthContext provides a way to access the authentication headers and fetch function
-// undefined means that the authentication has not been checked yet
-export const AuthContext = createContext<AuthContextValue | undefined>(
-  undefined
-);
-
-interface AuthProviderProps {
-  children?: React.ReactNode;
-}
-export const AuthProvider = ({ children }: AuthProviderProps) => {
-  const [auth, setAuth] = useState<AuthContextValue | undefined>(undefined);
-
-  useEffect(() => {
-    fetchOAuth()
-      .then((oAuthResult) => {
-        if (!oAuthResult) {
-          setAuth(createEmptyAuth());
-          return;
-        }
-        setAuth(createAuthFromOAuthResult(oAuthResult));
-      })
-      .catch((e: unknown) => {
-        console.error("Error fetching OAuth");
-        setAuth(createEmptyAuth());
-        console.error(e);
-      });
-  }, []);
-
-  return <AuthContext.Provider value={auth}>{children}</AuthContext.Provider>;
-};
